week2_workflow_orchestration/docker-requirements.txt
pandas==2.0.3
pyarrow==14.0.1
fastparquet==2023.10.1
prefect-aws==0.4.6
prefect-docker==0.4.4